+++ baRSS/Status Bar Menu/BarStatusItem.h
@@ -29,5 +29,6 @@
 - (void)setUnreadCountRelative:(NSInteger)count;
 - (void)asyncReloadUnreadCount;
 - (void)updateBarIcon;
+- (void)showWelcomeMessage;
 @end